[Version]: 1.0.0.6

+++ commands/main/verify.js
@@ -1,8 +1,8 @@
-const { member_role_id } = require('../../config.json')
+const { member_role_id, verify_role_id } = require('../../config.json')
 module.exports = {
 	description: 'verify the participant',
 	run (bot, message, args) {
-		if (message.member.roles.cache.has(member_role_id)) {
+		if (message.member.roles.cache.has(member_role_id) && message.member.roles.cache.has(verify_role_id)) {
 			message.channel.send('You are verified already.').then(msg => {
 				setTimeout(() => {
 					message.delete()
@@ -11,7 +11,7 @@ module.exports = {
 			})
 		} else {
 			message.member.roles.add(member_role_id)
-			message.member.roles.add(member_role_id)
+			message.member.roles.add(verify_role_id)
 			message.channel.send('You are now verified :)').then(msg => {
 				setTimeout(() => {
 					message.delete()
